requirements.txt
# Core dependencies
langchain
langchain-community
chromadb
pypdf
python-dotenv
pydantic
pydantic-settings

# Async HTTP client for Ollama integration
httpx

# API
fastapi
uvicorn[standard]

# Testing
pytest
pytest-asyncio
deepeval

# Utilities
python-multipart
pyyaml
requests
tqdm
google-generativeai

# Optional: Streamlit UI
# streamlit
